Start working on ToolTipController
